Sort functions out
