+++ AR_proto/os_combine/main.py
@@ -6,16 +6,15 @@ import time
 import datetime
 
 # import ar_module
-from vec_calc import find_vec
+from ar_module import Target, find_vec
 import motor
 import RPi.GPIO as GPIO
-import tarvec
 
 save_video = True
 
 
-# ar = ar_module.Ar_cansat()
-tg = tarvec.Target()
+# ar = Ar_cansat()
+tg = Target()
 
 
 # GPIO.setwarnings(False)
